+++ TenantScoreSheet/Models/Pets.cs
@@ -10,13 +10,13 @@
         public int? PetApprovedLandlordReferance2 { get; set; }
         public bool? NoOfCatsCompanion { get; set; }
         public int? NoOfCatsCompanions { get; set; }
-        public int? NoOfCatsCompanionPoints { get; set; }
+        public double? NoOfCatsCompanionPoints { get; set; }
         public bool? NoOfLargeDogsCompanion { get; set; }
         public int? NoOfLargeDogsCompanions { get; set; }
-        public int? NoOfLargeDogsCompanionPoints { get; set; }
+        public double? NoOfLargeDogsCompanionPoints { get; set; }
         public bool? NoOfSmallDogsCompanion { get; set; }
         public int? NoOfSmallDogsCompanions { get; set; }
-        public int? NoOfSmallDogsCompanionPoints { get; set; }       
+        public double? NoOfSmallDogsCompanionPoints { get; set; }       
         public string? CreatedBy { get; set; } = string.Empty!;
         public DateTime? CreatedDate { get; set; }
         public string? ModifiedBy { get; set; } = string.Empty!;
